pyproject.toml
[project]
name = "animation-retrieval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "dotenv>=0.9.9",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.49.0",
    "google-generativeai>=0.8.5",
    "huggingface>=0.0.1",
    "huggingface-hub>=1.0.1",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "torch>=2.9.0",
    "torchcodec>=0.8.1",
    "tqdm>=4.67.1",
    "tinytag>=1.6.0",
    "moviepy>=2.2.1",
]
